[fix] update feilds independantly and not found error

+++ src/student/student.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Student } from '../models/entities/student.entity';
@@ -10,13 +10,13 @@ export class StudentService {
   constructor(
     @InjectRepository(Student)
     private readonly studentRepository: Repository<Student>,
-  ) { }
+  ) {}
 
   // Create a new student
   async create(createStudentDto: CreateStudentDto): Promise<StudentEntity> {
     const student = this.studentRepository.create({
       common: {
-        ...createStudentDto
+        ...createStudentDto,
       },
     });
 
@@ -38,6 +38,7 @@ export class StudentService {
   // Find a student by ID
   async findById(id: number): Promise<StudentEntity> {
     const student = await this.studentRepository.findOneBy({ id });
+    if (!student) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
     const { common, ...rest } = student;
     return new StudentEntity({
       ...common,
@@ -51,7 +52,8 @@ export class StudentService {
         email,
       },
     });
-    if (!student) return null;
+    if (!student) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
+
     const { common, ...rest } = student;
     return new StudentEntity({
       ...common,
@@ -64,12 +66,18 @@ export class StudentService {
     id: number,
     updateStudentDto: Partial<CreateStudentDto>,
   ): Promise<StudentEntity | null> {
+    const student = await this.studentRepository.findOneBy({ id });
+    if (!student) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
     const result = await this.studentRepository.update(id, {
       common: {
-        ...updateStudentDto
+        name: updateStudentDto.name || student.common.name,
+        phoneNumber: updateStudentDto.phoneNumber || student.common.phoneNumber,
+        accessToken: updateStudentDto.accessToken || student.common.accessToken,
+        refreshToken:
+          updateStudentDto.refreshToken || student.common.accessToken,
       },
     });
-
+    console.log(result.affected);
     if (result.affected === 1) {
       const updatedStudent = await this.studentRepository.findOneBy({ id });
       if (updatedStudent) {
@@ -85,6 +93,8 @@ export class StudentService {
 
   // Delete a student by ID
   async remove(id: number): Promise<void> {
+    const student = await this.studentRepository.findOneBy({ id });
+    if (!student) throw new HttpException('Not Found', HttpStatus.NOT_FOUND);
     await this.studentRepository.delete(id);
   }
 }
